Threadify waiting for the slicer plugin
While the CuraEngine is a separate process, the python thread that is
supposed to be maintaining the Polar Cloud socket was stuck waiting for
sarge to wait for the external process to finish.  This change sticks it
on another thread so that the Polar Cloud will get to see the
"PREPARING" status.